under the snow

+++ 0x15-file_io/1-create_file.c
@@ -10,7 +10,7 @@
 int create_file(const char *filename, char *text_content)
 {
 	int fd;
-	ssize_t bytes_w;
+	int bytes_w;
 
 	if (filename == 0)
 		return (0);
